requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.10
joblib>=1.3

# BigQuery
google-cloud-bigquery>=3.15
google-cloud-bigquery-storage>=2.24
google-auth>=2.0

# Modeling / Analysis
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14

# Utilities
python-dotenv>=1.0
tqdm>=4.66